mitosis test

+++ eos/data.go
@@ -15,11 +15,11 @@ import (
 
 var lock sync.Mutex
 
-var launches = map[string]int{}
+var launches = map[string]int64{}
 
 // maxSessions is the maximum number of sessions during the lifetime of the node.
 // See Mitosis() for explanation.
-const maxSessions = 200
+const maxSessions = int64(200)
 
 // maxRuntime is the maximum time a node is available for launch requests
 const maxRuntime = 8 * time.Hour
